# ⚡ Optimized `HashNode.ToBencodex()`