Resolves #3922.

Sample benchmark with 1,000,000 `HashDigest<SHA256>`s:

|      Method |     Mean |   Error |   StdDev |       Gen0 |       Gen1 |     Gen2 | Allocated |
|------------ |---------:|--------:|---------:|-----------:|-----------:|---------:|----------:|
|   ByteArray | 214.1 ms | 4.20 ms |  7.14 ms | 12000.0000 |  4333.3333 | 666.6667 | 116.67 MB |
| ToByteArray | 417.9 ms | 9.98 ms | 28.97 ms | 29000.0000 | 10000.0000 |        - | 223.48 MB |